adcs-backend/requirements.txt
bcrypt<4.0.0
fastapi[standard]==0.135.1
passlib==1.7.4
psycopg2-binary==2.9.11
python-jose==3.5.0
sqlalchemy==2.0.48